refactor:InputLottoNumers를 Lotto로 통일
