+++ src/test/java/step2/domain/lotto/LottoWinningCheckerTest.java
@@ -28,4 +28,6 @@ class LottoWinningCheckerTest {
         assertThat(winningChecker).isNotNull();
 
     }
+
+
 }
