+++ src/test/java/step2/dto/LottoConfirmationRequestDtoTest.java
@@ -2,10 +2,8 @@ package step2.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step2.domain.InputLottoNumbers;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoConfirmationRequestDtoTest {
